# new_backend/api/openapi.yaml
openapi: 3.0.0
info:
  title: Swagger Studio Clock - OpenAPI 3.0
  description: Swagger for the Studio Clock application for MIEM.
  version: 1.0.0
servers:
  - url: https://clock.miem.tv
paths:
  /reconnect:
    post:
      summary: Connect Clocks to new OBS
      description: Connect Clocks to new OBS
      requestBody:
        description: Specify OBS credentials
        content:
          application/json:
            schema:
              required:
                - ip
                - port
              type: object 
              properties:
                ip:
                  type: string
                  example: 127.0.0.1
                port:
                  type: number
                  example: 4455
                password:
                  type: string
                  example: 1234
        required: true
      responses:
        "200":
          description: Successuful connection
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
        "400":
          description: Unsuccessful connection
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  description:
                    type: string
                    example: "Не удалось подключиться к OBS. Удостоверьтесь в правильности данных."
  /message:
    post:
      summary: Send message to director hint block
      description: Send message to director hint block
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  example: "Встаньте левее"
      responses:
        "200":
          description: Your body was delivered
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
        "400":
          description: The message is too long or cannot be converted to text
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  description:
                    type: string
                    example: "Сообщение слишком длинное (больше 26 символов)."
  /block/{operation}:
    post:
      summary: Start the block timer
      description: Start the block timer
      parameters:
        - name: operation
          in: path
          description: Operation for the block timer
          required: true
          schema:
            enum: [
              "start",
              "pause",
              "stop"
              ]
      responses:
        "200":
          description: Timer started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
        "400":
          description: Invalid operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  description:
                    type: string
                    example: "Такого события нет, вот список доступных: start, pause, stop"